deps: upd hnix to 0.14
Sorry for that the nice changelog stopped. With the number of upgrades I do to
the library internally & so changing the API which is wast - it become too
costly to maintain the level of a changelog I was maintaining.

I decided to adopt the strategy of noting only the most downstream-use-facing
changes (mostly Shorthands) & relying on the strength of Haskell API & me having
a strong Git commit history & commit documentation practices & providing
proactive support & helping people on requests.

What you use seems untouched by the 0.14 API changes.